Moving execution functions to exec.go
